refactor: replace login and keys commands
