+++ pkgs/standards/peagen/docs/secret_workflow_test.md
@@ -1,6 +1,6 @@
 # Peagen Secret Workflow Test
 
-This document records the steps to verify `peagen` CLI login, key, and secret management against a Postgres-backed gateway. It demonstrates that secrets are persisted to the database and retrieved by a worker.
+This document records the steps to verify `peagen` CLI public key, deploy key, and secret management against a Postgres-backed gateway. It demonstrates that secrets are persisted to the database and retrieved by a worker.
 
 ## Environment Setup
 
@@ -13,14 +13,14 @@ DQ_GATEWAY=http://127.0.0.1:8000/rpc uvicorn peagen.worker:app --host 127.0.0.1 
 
 A `.peagen.toml` referencing Redis and Postgres was placed in `/tmp/peagen_demo`.
 
-## Key Generation and Login
+## Key Generation and Public Key Upload
 
 ```bash
-peagen keys create --key-dir /tmp/peagen_demo/keys
-peagen login --gateway-url http://127.0.0.1:8000/rpc --key-dir /tmp/peagen_demo/keys
+peagen local deploykey create --key-dir /tmp/peagen_demo/keys
+peagen remote publickey upload --gateway-url http://127.0.0.1:8000/rpc --key-dir /tmp/peagen_demo/keys
 ```
 
-The login command succeeded, but `public_keys` remained empty as the gateway stores trusted keys in memory only.
+The upload command succeeded, but `public_keys` remained empty as the gateway stores trusted keys in memory only.
 
 ## Secret Management
 
@@ -42,4 +42,4 @@ The `secrets` table now contains the encrypted secret.
 
 ## Conclusion
 
-`peagen` successfully stored secrets in Postgres once the tenant existed. Workers retrieved the secret through the `remote secrets get` command. Public keys uploaded during `peagen login` were not persisted to Postgres in this setup.
+`peagen` successfully stored secrets in Postgres once the tenant existed. Workers retrieved the secret through the `remote secrets get` command. Public keys uploaded during `peagen remote publickey upload` were not persisted to Postgres in this setup.
