+++ pkgs/standards/peagen/peagen/cli/__init__.py
@@ -36,8 +36,8 @@ from peagen.cli.commands import (
     local_sort_app,
     local_template_sets_app,
     local_validate_app,
-    login_app,
-    keys_app,
+    local_publickey_app,
+    local_deploykey_app,
     local_secrets_app,
     show_app,
     remote_doe_app,
@@ -50,6 +50,8 @@ from peagen.cli.commands import (
     remote_task_app,
     remote_template_sets_app,
     remote_validate_app,
+    remote_publickey_app,
+    remote_deploykey_app,
     remote_secrets_app,
     local_analysis_app,
     remote_analysis_app,
@@ -190,8 +192,6 @@ def _global_remote_ctx(  # noqa: D401
 
 # ─────────────────────────── SUB-COMMAND REGISTRY ───────────────────────────
 
-app.add_typer(login_app)
-app.add_typer(keys_app, name="keys")
 app.add_typer(local_init_app, name="init")
 app.add_typer(fetch_app)
 app.add_typer(local_app, name="local")
@@ -213,6 +213,8 @@ local_app.add_typer(local_analysis_app)
 local_app.add_typer(local_template_sets_app, name="template-set")
 local_app.add_typer(local_validate_app)
 local_app.add_typer(local_secrets_app, name="secrets")
+local_app.add_typer(local_publickey_app, name="publickey")
+local_app.add_typer(local_deploykey_app, name="deploykey")
 local_app.add_typer(show_app, name="git")
 
 
@@ -228,6 +230,8 @@ remote_app.add_typer(remote_analysis_app, name="analysis")
 remote_app.add_typer(remote_template_sets_app, name="template-set")
 remote_app.add_typer(remote_validate_app)
 remote_app.add_typer(remote_secrets_app, name="secrets")
+remote_app.add_typer(remote_publickey_app, name="publickey")
+remote_app.add_typer(remote_deploykey_app, name="deploykey")
 
 if __name__ == "__main__":
     app()
